+++ analysis/HistogramDef.cc
@@ -19,6 +19,7 @@ along with BDSIM.  If not, see <http://www.gnu.org/licenses/>.
 #include "HistogramDef.hh"
 
 #include <iostream>
+#include <regex>
 #include <string>
 
 ClassImp(HistogramDef)
@@ -52,4 +53,10 @@ std::ostream& operator<< (std::ostream& out, const HistogramDef& s)
   std::string bs = s.GetBinningString();
   out << hs << " " << s.treeName << " " << s.histName << " " << bs << " " << s.variable << " " << s.selection << "\n";
   return out;
+}
+
+void HistogramDef::ReplaceStringInVariable(const std::string& match,
+                                           const std::string& replacement)
+{
+  variable = std::regex_replace(variable, std::regex(match), replacement);
 }