+++ analysis/HistogramDef.hh
@@ -56,6 +56,10 @@ public:
 
   /// Get the first string that defines the histogram in rebdsim for feedback.
   virtual std::string GetHistogramString() const = 0;
+
+  /// Little utility function.
+  void ReplaceStringInVariable(const std::string& match,
+                               const std::string& replacement);
   
   std::string treeName;
   std::string histName;